backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.27
python-dotenv==1.1.1
redis==6.2.0
pydantic==2.11.7
requests==2.32.4 